changed to compile on linux and added tla specs

+++ sample/include/latency_recorder.h
@@ -5,10 +5,12 @@
 #ifndef A_LESS_WRITEY_PAXOS_LATENCY_RECORDER_H
 #define A_LESS_WRITEY_PAXOS_LATENCY_RECORDER_H
 
+#include "stdint.h"
+
 struct latency_recorder;
 
 struct latency_recorder *
-latency_recorder_new(const char *output_file_path, struct timeval settle_in_time, u_int64_t latencies_to_record);
+latency_recorder_new(const char *output_file_path, struct timeval settle_in_time, uint32_t latencies_to_record);
 
 void latency_recorder_record(struct latency_recorder* recorder, unsigned long latency);
